TestingPython/requirements.txt
opencv-python
pyserial
RPLCD
Pillow
termcolor